Resolve #219: recreate symlinks when autoinstalling SDK
The SDK is unpacked using AntBuilder, and since Java is unable to create
symlinks, they are lost during the auto install procedure.

This fix attempts to unpack the .tgz or .tbz2 SDKs first by using `tar`.
If that fails (e.g. Windows), it resorts to the old method with
AntBuilder.